remove try_trait
The .ok_or() method on Option is enough for those two occasions that we
don't need to activate the complete try_trait feature just for a very
small and questionable ergonomics gain. This also allows us to more
easily discern which data exactly was invalid, as the .ok_or() in
different places could take different error values.

This also allows evtclib to be used on stable now.

Some noise is introduced in the diff due to automatically re-formatting
the source.