combine AgentKind and AgentName
Both of those are only used in lockstep anyway. If the AgentKind was
Player, the AgentName was also Player. Having the possibility of
mis-matched enum variants here was bad and always required an extra step
of "unwrap" that was not necessary.

This combination is the first step to simplify the handling of different
agent kinds.