force clippy to fail on warnings
There is "cargo clippy -- -D warnings", but that stops after the first
warning - and it displays them as an error. Not really the best
solution. The hacky workaround here is to make the script have a
non-zero return status when it found warnings, but still display them as
warnings only.