Cargo: only include sources
The source code itself is pretty small, however, the repository also
contains >40MB of test data in the form of evtc log files. We don't want
all consumers of this library to download all of this test data just
because they want to use the library.

This commit excludes everything except for the source (and some metadata
files) from the cargo package, allowing a thin distribution.

The git repository will still contain the tests, and the CI will still
run them.