ignore unknown statechanges in parse_events
The reason why we "unwrap" the error so late is because we want to
recover from this error, which means the file pointer has to be at the
right position. Unwrapping early would leave the pointer in the middle
of an event, which is not what we want.

If we want to bullet-proof this, it might be good to read the whole
event first into a buffer, and then read from that buffer instead.