re-introduce Boss
This is now the enum that contains the IDs of the single bosses, like
Nikare and Kenut. This means we can do away with the NIKARE_ID and such.

The enum is not publicly re-exported, as we re-export Encounter (which
is more of a replacement of the old Boss).

Special casing still remains (mostly in lib.rs), but we should be able
to do away with this now with a more general Encounter::bosses and
Boss::encounter methods.