Move game_mode to Encounter
If we already have an Encounter, it might be nice to determine the game
mode from it as well - without needing to go through the whole log
first.

This is especially useful for raidgrep, where we can use the early
filters - which don't have access to the whole Log item.