switch order of operands in SH analyzer
Comparing the int is a very cheap operation, and it is also a very good
indicator already that we've found the right event. Due to the short
circuiting behaviour of &&, it is better to check that first before
doing the Log::is_boss check - which is relatively costly. Remember that
we do this check for every buff application event!

This brings a speedup of around 50x:
new:    654.1±25.18µs
old:    34.9±0.69ms