speed up Log::is_boss
It is way cheaper to retrieve the correct agent first (especially with
the faster agent_by_addr) and then check if it is a boss, rather than
iterating over all bosses (which iterates through all characters), and
then iterate through the found bosses to check if any address matches.

The new code is a bit longer (and doesn't combine the functions as
nicely), but it is still readable and more performant - which is more
important.