fix dragonvoid ID
Dragonvoid is one of those weird bosses which consists of multiple
characters(?) that are structures and not even "normal" characters(??).
The first ID we used seemed to not catch the actual logs that are now
generated using an up-to-date arcDPS.

I'm not sure if the old ID was necessarily wrong, but for some reason,
it doesn't seem to match the actual ID that is currently used to log
this fight.